+++ CHANGELOG.md
@@ -14,6 +14,10 @@ This change log follows the format documented in [Keep a CHANGELOG].
 
 - Added `op-agent approve` command to pre-approve 1Password CLI commands.
 
+### Changed
+
+- `op-agent --version` (as well as `op-agent-client`) now prints only the version (e.g., `0.2.0`) instead of `op-agent version 0.2.0`.
+
 ## v0.1.0 - 2025-08-20
 
 Initial version